@@ -27,6 +27,7 @@
         "express-jwt": "^5.3.1",
         "graphql": "^0.13.2",
         "history": "^4.7.2",
+        "intersection-observer": "^0.5.0",
         "isomorphic-style-loader": "^4.0.0",
         "jquery": "3.3.1",
         "jsonwebtoken": "^8.2.0",
@@ -37,9 +38,9 @@
         "pretty-error": "^2.1.1",
         "prop-types": "^15.6.1",
         "query-string": "^6.0.0",
-        "react": "^16.2.0",
+        "react": "^16.5.1",
         "react-device-detect": "^1.5.9",
-        "react-dom": "^16.2.0",
+        "react-dom": "^16.5.1",
         "react-icons": "3.0.5",
         "react-lazyload": "^2.3.0",
         "react-redux": "^5.0.6",
