@@ -40,44 +40,36 @@ const trackedPlaylistIds = []; /** tracked the playlist/videos id both similar *
 class Home extends Component {
     state = {
       isDark: undefined,
-      isMenuOpen: false
+      isMenuOpen: false,
+      playlists: [],
+      videos: []
     }
 
-    componentWillReceiveProps(nextProps) {
-    	const {
-    		onHandleVideo,
-    		home: {
-    			playlists,
-    		},
-    	} = nextProps;
+    static getDerivedStateFromProps(nextProps, prevState) {
+      const {
+        onUpdatePlaylist,
+        onHandlePlaylist,
+        onHandleVideo,
+        home: {
+          playlists,
+        }
+      } = nextProps;
 
-    	if (
-    		playlists.meta.status === "success"
-    	) {
-    		playlists.data.map(playlist => {
-    			if (trackedPlaylistIds.indexOf(playlist.id) === -1) {
-    				trackedPlaylistIds.push(playlist.id);
-    				onHandleVideo(playlist);
-    			}
-	        });
-
-	        if (!activePlaylist) {
-	            activePlaylist = playlists.data[0];
-	            this.props.onUpdatePlaylist(activePlaylist.id);
-	        }
-	    }
-    }
-
-    componentWillMount() {
-    	const {
-    		onHandlePlaylist,
-    		home: {
-    			playlists,
-    		},
-    	} = this.props;
-    	if (playlists.meta.status !== 'success') {
-    		onHandlePlaylist();
-    	}
+      if (playlists.meta.status === 'loading' && prevState.playlists.length <= 0) {
+        onHandlePlaylist();
+      } else if (prevState.videos.length <= 0) {
+        playlists.data.map((playlist, index) => {
+          if (trackedPlaylistIds.indexOf(playlist.id) === -1) {
+            trackedPlaylistIds.push(playlist.id);
+            onHandleVideo(playlist);
+          }
+          if (!activePlaylist && index === 0) {
+            activePlaylist = playlists.data[0];
+            onUpdatePlaylist(activePlaylist.id);
+          }
+        });
+      }
+      return { ...prevState, playlists };
     }
 
     componentDidMount() {
@@ -221,14 +213,14 @@ class Home extends Component {
 
     handleSlideNext = (scrollIndex = 0) => {
       this.sliderRefs.sort((a, b) => a.sortOrder - b.sortOrder);
-      if (this.sliderRefs[scrollIndex]) {
+      if (this.sliderRefs[scrollIndex] && this.sliderRefs[scrollIndex].slickNext) {
         this.sliderRefs[scrollIndex].slickNext();
       }
     }
 
     handleSlidePrev = (scrollIndex = 0) => {
       this.sliderRefs.sort((a, b) => a.sortOrder - b.sortOrder);
-      if (this.sliderRefs[scrollIndex]) {
+      if (this.sliderRefs[scrollIndex] && this.sliderRefs[scrollIndex].slickPrev) {
         this.sliderRefs[scrollIndex].slickPrev();
       }
     }
@@ -263,11 +255,11 @@ class Home extends Component {
       return (
         <div>
           <Header libraryOff className={styles.placeholder__header} isDark={isDark} isMobile />
-          {status === 'loading' && <HomePlaceholder />}
-          {status === 'error' &&
-					<div className={styles.home__error_container}>Ada Error kawan: {error}</div>
+          {status === 'loading' || videos.meta.status === 'loading' && <HomePlaceholder />}
+          {status === 'error' || videos.meta.status === 'error' &&
+					<div className={styles.home__error_container}>Ada Error kawan: {error || "failed to load API Videos NOT FOUND"}</div>
           }
-          {status === 'success' &&
+          {status === 'success' && videos.meta.status === 'success' &&
                     <div>
                       <HomeMobileMenu
                         isDark={isDark}
@@ -291,6 +283,7 @@ class Home extends Component {
             status === 'success'
 					&& videos
                     && videos.data.length > 0
+                    && videos.data.length === playlists.data.length
                     && videos.data.map(video => {
                     	const { id, sortOrder } = video.meta;
                     	return (
@@ -347,10 +340,10 @@ class Home extends Component {
                                     >
                                       <LazyLoadBeta src={layer3}
                                         containerClassName={styles.home__parallax_layer_3_info}
-                                        style={{ color: isDark ? "black" : "white" }}
                                         alt=""
                                       >
-                                        <div className={styles.home__parallax_layer_3_detail}>
+                                        <div className={styles.home__parallax_layer_3_detail}
+                                          style={{ color: isDark ? "black" : "white" }}>
                                           <h4
                                             className={styles.home__parallax_layer_3_title}
                                           >
