@@ -1,5 +1,5 @@
 import { get } from 'axios'
-import { HOME_PLAYLIST_ENDPOINT, HISTORY_ENDPOINT, VIDEOS_ENDPOINT } from './endpoints'
+import { HOME_PLAYLIST_ENDPOINT, HISTORY_ENDPOINT, SEARCH_VIDEOS_ENDPOINT } from './endpoints'
 import utils from './util'
 
 const getHomePlaylist = (payload) => {
@@ -29,13 +29,19 @@ const getHomeVideo = ({ id }) => {
   return get(`${HOME_PLAYLIST_ENDPOINT}/${id}`).then(
     (response) => {
       const result = utils.normalizeHomeVideo(response)
-      return [...result[0]] || []
+      return {
+        meta: {
+          status: 'success',
+          error: '',
+        },
+        data: [...result[0]] || [],
+      }
     }
   ).catch((error) => {
     return {
       meta: {
         status: 'error',
-        text: `home/getHomeVideo ~ ${error}`,
+        error: `home/getHomeVideo ~ ${error}`,
       },
       data: [],
     }
@@ -46,7 +52,6 @@ const getAllHistory = (payload) => {
   return get(`${HISTORY_ENDPOINT}`, { ...payload }).then(
     (response) => {
       const result = utils.normalizeHistory(response)
-      // console.log('RES', result)
       return {
         meta: {
           status: result.length > 0 ? 'success' : 'no_result',
@@ -66,7 +71,7 @@ const getAllHistory = (payload) => {
 }
 
 const getSearchVideo = (payload) => {
-  return get(`${VIDEOS_ENDPOINT}`, { ...payload }).then(
+  return get(`${SEARCH_VIDEOS_ENDPOINT}`, { ...payload }).then(
     (response) => {
       const result = utils.normalizeSearchVideo(response);
       return {
@@ -81,7 +86,7 @@ const getSearchVideo = (payload) => {
     return {
       meta: {
         status: 'error',
-        error: `video/getSearchVideo ~ ${error}`,
+        error: `search/getSearchVideo ~ ${error}`,
       },
       data: [],
     }
