@@ -41,47 +41,51 @@ class Header extends Component {
       backButtonOn = false,
       isMobile = false,
       title = '',
-    } = this.props
+    } = this.props;
     const color = isDark ? 'black' : 'white';
-
+    const logoDark = isDark ? true : false;
     return (
       <div className={styles.header__container}>
         <div className={styles.header__logo_wrapper}>
           {!logoOff &&
             <Link to="/">
-              {isDark
-                ? <LazyLoadBeta src={isMobile ? logoLandscapeBlue : logoBlue} containerClassName={styles.header__logo}/>
-                : <LazyLoadBeta src={isMobile ? logoLandscapeGrey : logoGrey} containerClassName={styles.header__logo}/>
+              {
+                logoDark &&
+                <LazyLoadBeta src={isMobile ? logoLandscapeBlue : logoBlue} containerClassName={styles.header__logo}/>
+              }
+
+              {
+                !logoDark && <LazyLoadBeta src={isMobile ? logoLandscapeGrey : logoGrey} containerClassName={styles.header__logo}/>
               }
             </Link>
           }
           {backButtonOn && (
-            <LazyLoad>
+            <LazyLoadBeta>
               <div className={styles.header__back_button} onClick={this.handleGoBack}>
                 <IoIosArrowRoundBack size={32} color={color} />
               </div>
-            </LazyLoad>
-          )}
-        </div>
-        <div className={styles.header__library_wrapper} style={{ color }}>
-          {!libraryOff && (
-            <LazyLoadBeta>
-              <Link
-                className={styles.header__library_link_wrapper}
-                to="/movie-library"
-                style={{ color }}
-              >
-                <span
-                  className={styles[`header__library_logo_${color}`]}
-                  alt="library"
-                  style={{ width: '32px', height: '32px', }}
-                />
-              </Link>
             </LazyLoadBeta>
           )}
-          {title}
         </div>
-        {!rightMenuOff && <RightMenu color={color} searchOff={searchOff} />}
+        {/* <div className={styles.header__library_wrapper}> */}
+        {!libraryOff && (
+          <LazyLoadBeta>
+            <Link
+              className={styles.header__library_link_wrapper}
+              to="/movie-library"
+              style={{ color }}
+            >
+              <span
+                className={styles[`header__library_logo_${color}`]}
+                alt="library"
+                style={{ width: '32px', height: '32px', }}
+              />
+            </Link>
+          </LazyLoadBeta>
+        )}
+        {title}
+        {/* </div> */}
+        {!rightMenuOff && <RightMenu color={color} searchOff={searchOff}/>}
       </div>
     )
   }
