@@ -3,12 +3,12 @@ import user from './user';
 import runtime from './runtime';
 import home from './home';
 import history from './history';
-import video from './video'
+import search from './search'
 
 export default combineReducers({
   user,
   runtime,
   home,
   history,
-  video
+  search
 });
